chore: rename table components using kebab-case
